Add test for rest-client reactive timeout
Follow up on: #19223
